+++ src/PSRule.Rules.Azure/rules/Azure.Common.Rule.ps1
@@ -7,9 +7,8 @@ function global:ResourceType {
     param (
         [String]$ResourceType
     )
-
     process {
-        return $TargetObject.ResourceType -eq $ResourceType;
+        return $Rule.TargetType -eq $ResourceType;
     }
 }
 
@@ -40,7 +39,7 @@ function global:HasPeerNetwork {
     [OutputType([System.Boolean])]
     param ()
     process {
-        if ($TargetObject.ResourceType -ne 'Microsoft.Network/virtualNetworks') {
+        if ($Rule.TargetType -ne 'Microsoft.Network/virtualNetworks') {
             return $False;
         }
         $peers = $TargetObject.Properties.virtualNetworkPeerings;
@@ -62,7 +61,7 @@ function global:GetOrderedNSGRules {
 
 function global:SupportsAcceleratedNetworking {
     process {
-        if ($TargetObject.ResourceType -ne 'Microsoft.Compute/virtualMachines' -or !(IsExport)) {
+        if ($Rule.TargetType -ne 'Microsoft.Compute/virtualMachines' -or !(IsExport)) {
             return $False;
         }
 
@@ -107,7 +106,7 @@ function global:IsAppGwPublic {
     [OutputType([System.Boolean])]
     param ()
     process {
-        if ($TargetObject.ResourceType -ne 'Microsoft.Network/applicationGateways') {
+        if ($Rule.TargetType -ne 'Microsoft.Network/applicationGateways') {
             return $False;
         }
 
@@ -128,7 +127,7 @@ function global:IsAppGwWAF {
     [OutputType([System.Boolean])]
     param ()
     process {
-        if ($TargetObject.ResourceType -ne 'Microsoft.Network/applicationGateways') {
+        if ($Rule.TargetType -ne 'Microsoft.Network/applicationGateways') {
             return $False;
         }
         if ($TargetObject.Properties.sku.tier -notin ('WAF', 'WAF_v2')) {
@@ -146,7 +145,7 @@ function global:IsWindowsOS {
     [OutputType([System.Boolean])]
     param ()
     process {
-        if ($TargetObject.ResourceType -ne 'Microsoft.Compute/virtualMachines') {
+        if ($Rule.TargetType -ne 'Microsoft.Compute/virtualMachines') {
             return $False;
         }
         return $TargetObject.Properties.storageProfile.osDisk.osType -eq 'Windows';
@@ -158,7 +157,7 @@ function global:IsLinuxOS {
     [OutputType([System.Boolean])]
     param ()
     process {
-        if ($TargetObject.ResourceType -ne 'Microsoft.Compute/virtualMachines') {
+        if ($Rule.TargetType -ne 'Microsoft.Compute/virtualMachines') {
             return $False;
         }
         return $TargetObject.Properties.storageProfile.osDisk.osType -eq 'Linux';
@@ -172,16 +171,16 @@ function global:SupportsTags {
     param ()
     process {
         if (
-            ($TargetObject.ResourceType -eq 'Microsoft.Subscription') -or
-            ($TargetObject.ResourceType -like 'Microsoft.Authorization/*') -or
-            ($TargetObject.ResourceType -like 'Microsoft.Billing/*') -or
-            ($TargetObject.ResourceType -like 'Microsoft.Classic*') -or
-            ($TargetObject.ResourceType -like 'Microsoft.Consumption/*') -or
-            ($TargetObject.ResourceType -like 'Microsoft.Gallery/*') -or
-            ($TargetObject.ResourceType -like 'Microsoft.Resources/*') -or
-            ($TargetObject.ResourceType -like 'Microsoft.Security/*') -or
-            ($TargetObject.ResourceType -like 'microsoft.support/*') -or
-            ($TargetObject.ResourceType -like 'Microsoft.WorkloadMonitor/*')
+            ($Rule.TargetType -eq 'Microsoft.Subscription') -or
+            ($Rule.TargetType -like 'Microsoft.Authorization/*') -or
+            ($Rule.TargetType -like 'Microsoft.Billing/*') -or
+            ($Rule.TargetType -like 'Microsoft.Classic*') -or
+            ($Rule.TargetType -like 'Microsoft.Consumption/*') -or
+            ($Rule.TargetType -like 'Microsoft.Gallery/*') -or
+            ($Rule.TargetType -like 'Microsoft.Resources/*') -or
+            ($Rule.TargetType -like 'Microsoft.Security/*') -or
+            ($Rule.TargetType -like 'microsoft.support/*') -or
+            ($Rule.TargetType -like 'Microsoft.WorkloadMonitor/*')
         ) {
             return $False;
         }
@@ -197,9 +196,9 @@ function global:SupportsRegions {
     param ()
     process {
         if (
-            ($TargetObject.ResourceType -eq 'Microsoft.Subscription') -or
-            ($TargetObject.ResourceType -eq 'Microsoft.AzureActiveDirectory/b2cDirectories') -or
-            ($TargetObject.ResourceType -eq 'Microsoft.Network/trafficManagerProfiles') -or
+            ($Rule.TargetType -eq 'Microsoft.Subscription') -or
+            ($Rule.TargetType -eq 'Microsoft.AzureActiveDirectory/b2cDirectories') -or
+            ($Rule.TargetType -eq 'Microsoft.Network/trafficManagerProfiles') -or
             ($TargetObject.Location -eq 'global')
         ) {
             return $False;
@@ -333,7 +332,7 @@ function global:WithinCIDR {
 # Determine if the VM is using a promo SKU.
 function global:IsVMPromoSku {
     process {
-        if ($TargetObject.ResourceType -ne 'Microsoft.Compute/virtualMachines') {
+        if ($Rule.TargetType -ne 'Microsoft.Compute/virtualMachines') {
             return $False;
         }
         return $TargetObject.Properties.hardwareProfile.vmSize -like '*_Promo';